Land Recent QUIC Changes.
Remove InterArrivalSender and Receiver from QUIC because they are never
used and increase maintenance.

Merge internal change: 65742790
https://codereview.chromium.org/256393002/


Remove support for the Public Reset packets of QUIC_VERSION_13.

Merge internal change: 65740152
https://codereview.chromium.org/256383002/


Remove WillAcceptStreamFrames codepath since we effectively don't use
this code, and flow control is the right way to deal with this issue anyways.

Removes currently implemented protection measure in
<internal server>/quic to use max_frame_memory_ as the maximum stream
buffer size. This limit (i) is currently set to max_size_t and is
therefore not currently really useful, and (ii) will be more correctly
replaced by flow control. (See: b/14235094)

Merge internal change: 65737468
https://codereview.chromium.org/260233004/


Test change: Set return types of response_body_size() and
response_size() in SimpleClient to int64 to make them large enough to
hold number of response bytes.

Merge internal change: 65728046
https://codereview.chromium.org/254013004/


Swapping packet writer wrappers back to packet writers in preparation
for internal server QoS changes.  Not flag protected.

Using Quic packet writers directly instead of wrapping them.

Merge internal change: 65720400
https://codereview.chromium.org/250843009/


Only send a QUIC BLOCKED frame once at a given "sent bytes" offset.

Merge internal change: 65654681
https://codereview.chromium.org/251903007/


Add Kathleen Nichols min rtt algorithm to QUIC's RttStats, so it can be
used with BBR.

Merge internal change: 65596807
https://codereview.chromium.org/257133002/


Discard all unencrypted crypto packets in QUIC once the connection moves
to forward secure.

Merge internal change: 65592364
https://codereview.chromium.org/260243003/


Remove unneccessary .get() calls to scoped_ptr (reported by clang-tidy).

Merge internal change: 65590426
https://codereview.chromium.org/260033005/


Sync'ing with internal source code.

+ Deleted a comment per review comments from wtc.

Merge internal change: 65565864
https://codereview.chromium.org/258063011/


Change QUIC SendAlgorithmInterface's UpdateRtt method to OnRttUpdated
and supply a largest observed to the send algorithm instead of the
latest rtt sample argument.

Necessary for the new BBR implementation, which needs to know what
packet was acked when an rtt sample is taken.

Merge internal change: 65541913
https://codereview.chromium.org/258223002/


Print error code as string on write failure in QuicConnection.

Merge internal change: 65519182
https://codereview.chromium.org/251143006/


R=rch@chromium.org

Review URL: https://codereview.chromium.org/263473003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267358 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 82168ee207dba9beee9f73d8e6543740ef84c06b

+++ net/tools/quic/test_tools/quic_test_client.cc
@@ -284,7 +284,7 @@ int QuicTestClient::response_header_size() const {
   return response_header_size_;
 }
 
-int QuicTestClient::response_body_size() const {
+int64 QuicTestClient::response_body_size() const {
   return response_body_size_;
 }
 
@@ -468,7 +468,7 @@ const BalsaHeaders* QuicTestClient::response_headers() const {
   }
 }
 
-int QuicTestClient::response_size() const {
+int64 QuicTestClient::response_size() const {
   return bytes_read_;
 }
