+++ net/tools/quic/test_tools/simple_client.cc
@@ -29,7 +29,7 @@ int SimpleClient::HalfClose() {
 
 int SimpleClient::response_header_size() const { return 0; }
 
-int SimpleClient::response_body_size() const { return 0; }
+int64 SimpleClient::response_body_size() const { return 0; }
 
 }  // namespace net
 }  // namespace tools
